third_party/WebKit/LayoutTests/svg/custom/intersection-list-null-expected.txt

CONSOLE ERROR: line 8: Uncaught TypeError: Failed to execute 'checkIntersection' on 'SVGSVGElement': parameter 1 is not of type 'SVGElement'.
PASS
